Changes to be committed; modified:   main.py; modified:   rectoVerso.py;
